% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_file_content.R
\name{print_file_content}
\alias{print_file_content}
\title{A function which reads the text from a file and prints it}
\usage{
print_file_content(file_path)
}
\arguments{
\item{file_path}{character - path to the file}
}
\description{
A function which reads the text from a file and prints it
}
\examples{
path <- system.file("exdata", "file.txt",
                    package = "pkgTemplateR")
print_file_content(file_path = path)
}
